refactor: split oversized file per 800-line cap requirement
